Consolidate interfaces used in login process into folder
