+++ src/app/services/login/dto/account-details.ts
@@ -1,5 +1,5 @@
-import { UserType } from "../../store/user-type.enum";
-import { RiderType } from "../../store/rider-type.enum";
+import { UserType } from "../../../store/user-type.enum";
+import { RiderType } from "../../../store/rider-type.enum";
 
 export interface AccountDetails {
   username: string;
