+++ src/app/services/login/login.service.ts
@@ -3,9 +3,9 @@ import { User } from 'src/app/users/user';
 import { HttpClient } from "@angular/common/http";
 import { UserType } from "../../store/user-type.enum";
 import { RiderType } from "../../store/rider-type.enum";
-import { AccountDetails } from "./account-details";
+import { AccountDetails } from "./dto/account-details";
 import { Util } from "../../users/util";
-import { Credentials } from "./credentials";
+import { Credentials } from "./dto/credentials";
 
 @Injectable({
   providedIn: 'root'
